Fixes file system watcher within YamlDataStore
The file system watcher is meant to watch a directory and takes such as its single parameter. To watch a single file within said directory, the filter (second param) must be the file name.

Exception is thrown without this fix as the path given is a file, not a directory.