Fixes NuGetPackageManager installing out of range versions
This occurs when the version range as an exclusive bound. The package manager previously would attempt to install the upper bound as it had no check for whether or not the bound was exclusive rather than inclusive.

By getting all package versions, the package manager will always install the best matched version.