Fix lag on player connect & disconnect
The events `OnPlayerConnected`, `OnPlayerDisconnected`, `OnRejectingPlayer`, and `OnPendingPlayerConnecting` get the user information on the main thread and stall the server. This isn't noticeable if you only have a few users in your openmod.users.yaml file, but when there are thousands of users in the file, then it will cause a massive lag spike whenever someone joins or leaves.

I've changed the event code to run on a different thread using AsyncHelper.Schedule(), and it's been working great on my servers.